Refactor CloudinaryField: move upload options and validation to form
